[#88] Manager inquiry
Done
1. Manager inquiry api

+++ web/back/src/main/java/com/neo/back/infoManaging/service/UserInquiryService.java
@@ -9,6 +9,7 @@ import com.neo.back.infoManaging.dto.UserInquiryListDto;
 import com.neo.back.infoManaging.dto.UserInquiryToAnswer;
 import com.neo.back.infoManaging.dto.UserPostInquiryDto;
 import com.neo.back.infoManaging.entity.UserInquiry;
+import com.neo.back.infoManaging.middleware.RootAPI;
 import com.neo.back.infoManaging.repository.UserInquiryRepository;
 import com.neo.back.service.dto.ServerListDto;
 
@@ -24,6 +25,7 @@ import java.time.format.DateTimeFormatter;
 @RequiredArgsConstructor
 public class UserInquiryService {
     private final UserInquiryRepository userInquiryRepo;
+    private final RootAPI rootAPI;
 
     public Mono<Object> postUserInquiry(User user,UserPostInquiryDto inquiryData){
         UserInquiry inquiry = new UserInquiry();
@@ -43,7 +45,7 @@ public class UserInquiryService {
         Optional<UserInquiry> inquirys = userInquiryRepo.findById(inquiryId);
         if(inquirys.isPresent()){
             UserInquiry inquiry = inquirys.get();
-
+            if(!this.rootAPI.checkUser(user,inquiry)) return Mono.just("not your inquiry");
             return Mono.just(new UserInquiryToAnswer(
                 inquiry.getId(),
                 inquiry.getAnswerOrNot(),
@@ -73,6 +75,8 @@ public class UserInquiryService {
     public Mono<String> deleteUserInquiry(User user, Long inquiryId){
         Optional<UserInquiry> inquirys = userInquiryRepo.findById(inquiryId);
         if(inquirys.isPresent()){
+            UserInquiry inquiry = inquirys.get();
+            if(!this.rootAPI.checkUser(user,inquiry)) return Mono.just("not your inquiry");
             userInquiryRepo.deleteById(inquiryId);
             return Mono.just("success to delete Inquiry");
         }
@@ -82,12 +86,14 @@ public class UserInquiryService {
     }
 
     public Mono<Object> postManagerInquiry(User user,ManagerPostInquiryDto inquiryData){
-        Optional<UserInquiry> inquirys = userInquiryRepo.findById(inquiryData.getId());
+        if(!this.rootAPI.checkManager(user)) return Mono.just("not Manager");
+        Optional<UserInquiry> inquirys = userInquiryRepo.findById(inquiryData.getInquiryId());
         if(inquirys.isPresent()){
             UserInquiry inquiry = inquirys.get();
             String answerTime = LocalDateTime.now().format(DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss"));
             inquiry.setManagerAnswer(inquiryData.getInquiry());
             inquiry.setManagerAnswerDate(answerTime);
+            inquiry.setAnswerOrNot(true);
             userInquiryRepo.save(inquiry);
             return Mono.just("success to answer userInquiry");
         }
@@ -97,6 +103,7 @@ public class UserInquiryService {
     }
 
     public Mono<Object> getManagerInquiryToAnswer(User user, Long inquiryId){
+        if(!this.rootAPI.checkManager(user)) return Mono.just("not Manager");
         Optional<UserInquiry> inquirys = userInquiryRepo.findById(inquiryId);
         if(inquirys.isPresent()){
             UserInquiry inquiry = inquirys.get();
@@ -112,7 +119,32 @@ public class UserInquiryService {
             ));
         }
         else{
-            return Mono.just("fail to answer userInquiry");
+            return Mono.just("fail to get ManagerInquiry");
         }
     }
+
+    public Mono<Object> getManagerInquiryList(User user){
+        if(!this.rootAPI.checkManager(user)) return Mono.just("not Manager");
+        List<UserInquiry> inquirys = userInquiryRepo.findAll();
+        return Mono.just(inquirys.stream()
+        .map(inquiry -> new UserInquiryListDto(
+         inquiry.getId(),
+         inquiry.getAnswerOrNot(),
+         inquiry.getUserInquiryTitle(),
+         inquiry.getUserInquiryDate()))
+         .collect(Collectors.toList()));
+    }
+
+    public Mono<String> deleteManagerInquiry(User user, Long inquiryId){
+        if(!this.rootAPI.checkManager(user)) return Mono.just("not Manager");
+        Optional<UserInquiry> inquirys = userInquiryRepo.findById(inquiryId);
+        if(inquirys.isPresent()){
+            userInquiryRepo.deleteById(inquiryId);
+            return Mono.just("success to delete Inquiry by manager");
+        }
+        else{
+            return Mono.just("fail to delete Inquiry by manager");
+        }
+    }
+
 }
